requirements.txt
axelrod
jupyter
matplotlib
mystmd
numpy
python-ternary
